requirements.txt
transformers
datasets
torch
pillow
google-cloud-documentai
google-genai
requests
pydantic
fastapi